[eclipse-iceoryx#956] Fix datatype to avoid conversion errors

+++ examples/cxx/request_response_dynamic_data/src/client.cpp
@@ -11,6 +11,7 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 #include "iox2/iceoryx2.hpp"
+#include <cstdint>
 
 constexpr iox::units::Duration CYCLE_TIME = iox::units::Duration::fromSeconds(1);
 
@@ -43,8 +44,9 @@ auto main() -> int {
     while (true) {
         auto required_memory_size = std::min(1000000, counter * counter); // NOLINT
         auto request = client.loan_slice_uninit(static_cast<uint64_t>(required_memory_size)).expect("loan successful");
-        auto initialized_request = request.write_from_fn(
-            [&](auto byte_idx) { return (byte_idx + static_cast<uint64_t>(counter)) % 255; }); // NOLINT
+        auto initialized_request = request.write_from_fn([&](auto byte_idx) {
+            return static_cast<uint8_t>((byte_idx + static_cast<uint64_t>(counter)) % 255);
+        }); // NOLINT
         auto pending_response = send(std::move(initialized_request)).expect("send successful");
         std::cout << "send request " << counter << " with " << required_memory_size << " bytes ..." << std::endl;
